Destriper equations
I've modified the destriper equations. Now, it is
more similar at the gls approach. The alpha and knee frequencies
are tuned for every month/day/detectors

I have to implement the `bin_map` function and the `denoise`
function.

The denoise function has to use the fftw methods. Use the C2R.